# app/govern/docs/swagger.yaml
basePath: /api/v1
definitions:
  bll.AddApiReq:
    properties:
      enable:
        type: boolean
      group:
        type: string
      method:
        type: string
      path:
        type: string
    required:
    - method
    - path
    type: object
...
  bll.EnableApiReq:
    properties:
      enable:
        type: boolean
    required:
    - enable
    type: object
...
  bll.ListApiResp:
    properties:
      data:
        type: object
      total:
        type: integer
    type: object
...
  bll.ProfileApiResp:
    properties:
      createdAt:
        type: integer
      creator:
        type: string
      enable:
        type: boolean
      group:
        type: string
      method:
        type: string
      path:
        type: string
      updatedAt:
        type: integer
    type: object
...
  bll.UpdateApiReq:
    properties:
      group:
        type: string
      method:
        type: string
      path:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: sfshf@github.com
    name: gavin
    url: http://github.com/sfshf
  description: This is a back-end government app.
  license:
    name: MIT
    url: https://github.com/sfshf/sprout/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Govern APIs
  version: 0.0.1-beta
paths:
  /apis:
    get:
      description: Get a list of api.
      operationId: api-list
      parameters:
      - in: query
        name: createdAtBegin
        type: integer
      - in: query
        name: createdAtEnd
        type: integer
      - in: query
        name: creator
        type: string
      - in: query
        name: enable
        type: boolean
      - in: query
        name: group
        type: string
      - in: query
        name: method
        type: string
      - in: query
        name: orderBy
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: path
        type: string
      - in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: api list.
          schema:
            $ref: '#/definitions/bll.ListApiResp'
      security:
      - ApiKeyAuth: []
      summary: Get a list of api.
      tags:
      - api
    post:
      consumes:
      - application/json
      description: Add a new api.
      operationId: api-add
      parameters:
      - description: required attributes to add a new api.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.AddApiReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Add a new api.
      tags:
      - api
  /apis/:id:
    delete:
      description: Evict a specific api.
      operationId: api-evict
      parameters:
      - description: id of the api to evict.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Evict a specific api.
      tags:
      - api
    get:
      description: Get the profile of a api.
      operationId: api-profile
      parameters:
      - description: id of the api.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: profile of the api.
          schema:
            $ref: '#/definitions/bll.ProfileApiResp'
      security:
      - ApiKeyAuth: []
      summary: Get infos of an api.
      tags:
      - api
    put:
      consumes:
      - application/json
      description: Update a specific api.
      operationId: api-update
      parameters:
      - description: id of the api to update.
        in: path
        name: id
        required: true
        type: string
      - description: attributes need to update.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.UpdateApiReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Update a specific api.
      tags:
      - api
  /apis/:id/enable:
    patch:
      consumes:
      - application/json
      description: Enable or disable an api.
      operationId: api-enable
      parameters:
      - description: id of the api.
        in: path
        name: id
        required: true
        type: string
      - description: true for enable, or false for disable.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.EnableApiReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Enable or disable an api.
      tags:
      - api
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"